Add test for disjunction choices
We add a new test for disjunction choices which uses our custom surface
syntax. It follows the example in the blog post announcing disjunction
choices, which is still under review.

We also remove the old tests for flexible controllers. The test for
DAML-LF 1.1 is not run anymore anyway and the test for DAML-LF 1.2 does not
use our custom surface syntax but the type class implementation underlying
it.