Add CODEOWNERS with lenient permissions for incremental refinement.
We decided to set individual GitHub handles in the CODEOWNERS file and
to do so by letting people claim ownership in individual PRs. Thus, this
commit contains an initial version of the CODEOWNERS file that guards
that file, but leaves the full team with access to other files
initially.

Not all files in the root are explicitly listed, but they can be added
if someone wants to claim ownership of them.