Implement TextMap.toList (2nd attempt) (#508)
Implement the compiler logic needed to expose DAML-LF's `MAP_TO_LIST` as
`TextMap.toList` in the surface language.

This is part of #409. The type definition and module need to be put in the
right place before I consider this done. Once it's done, I'll add it to the
release notes.