Inline record projections on this in maintainer expression
If we don't inline let bindings of the form `x = this.f`, the rewriting of
the maintainer expression in terms of the key can fail when it shouldn't.
Usually, these bindings do not exist in the surface language but are produced
by the compiler. I think the performance penalty we might pay for this is
negligible.

This fixes #384.