daml-integration-kit: release PRE-ALPHA version in SDK (#440)
The documentation is in a state that interested third-parties can
consume; and will likely find valuable to consume. It also contains a
description of the current status of the integration kit and the roadmap
outlining its expected evolution.

We have added links to GitHub issues where documentation is missing, or
where it refers to missing pieces of code.