Return transaction in SubmitAndWaitResponse
When submitting a command via the CommandService (SubmitAndWait), it is
fairly cumbersome to get hold of the resulting transaction. Since the
response (Empty) doesn't contain any offset, transactionId, or other
metadata, the user has go wade through the transaction stream and check
for the transaction with the right commandId. This is not a good user
experience.

The SubmitAndWait method now always returns the transaction id.
The new method SubmitAndWaitForTransaction returns the flat transaction,
and the new method SubmitAndWaitForTransactionTree returns the
transaction tree.

Fixes #406.