Fix location information
The location information in `LArg` is meant to be for the function the
argument is applied to. Hence the first change was wrong. Since
`withTmArg` is not particularly useful when you want to apply the
expression passed into the continuation using `EApp`, we drop the
location information on the floor.