Avoid producing a few useless lambdas during conversion to DAML-LF (#866)
* Avoid producing a few useless lambdas during conversion to DAML-LF

This also removes a bit of code duplication by factoring out a common
pattern.

There are far more places where this technique can be applied. I'll do so
in a follow up PR.

* Fix location information

The location information in `LArg` is meant to be for the function the
argument is applied to. Hence the first change was wrong. Since
`withTmArg` is not particularly useful when you want to apply the
expression passed into the continuation using `EApp`, we drop the
location information on the floor.

* Lift withTyArg and withTmArg to the top level

* Fix type error