add non-strict mergify option
Add the option of setting the `automerge-dangerous-but-fast` label on a
PR to request mergify to merge it ASAP, i.e. as soon as tests pass,
skipping the important "rebase" step, making it prone to "race
condition" issues getting introduced and breaking master. This should
only be used for either very safe or very urgent work.

Please keep in mind `master` is the stable, shared branch from which
everyone works and its primary goal is to be stable. It is not the only
way to share code and should not be the primary way to share code within
small groups of people working on the same thing together.