Fold "damlc package-new" into "damlc build" and hide "damlc compile" (#1291)
The only difference between `damlc package-new` and `damlc build` was
that `package-new` did not initialize the package database. This
commit adds a `--init-package-db=yes/no/auto` flag to control
initialization of the package database.

I’ve also hidden compile since it should only be used for testing and
debugging purposes.