Test damlc against all supported DAML-LF version (#1337)
I disabled this in the past because the tests were to slow. They have gotten
quite a bit faster since then and I'm slowly getting scared by the fact we
don't do this right now. Also, it incentivices us to drop support for old
DAML-LF versions more frequently, which will most of the time also remove
complexity from the compiler.