Rewrite coercion conversion in higher-order abstract syntax style
The current version in first-order abstract syntax style does not play well
with dictionary sanitization in the corner case of a type class which has
exactly on method and this method has a default implementation. As a side
effect we also eliminate a few lambdas in the generated code.

I've deliberatly removed the handling of forall coercions and coercion
applications since we don't have any code triggering them. I'll add them
back as soon as we find such code. In the meantime, I prefer not handling
these cases and blowing up loudly rather than silently doing something
which might be wrong.