Detect type class instances during template resugaring qualified
Currently, we match on any class named `Template`. This PR changes it such
that we only matchon `Template` from `DA.Internal.Template`. Similarly for
`Choice` and `TemplateKey`.

Unfortunately, default method implementation names do not carry a module
name with them, this we need to continue matching on them unqualified.

This change is a preparation for #1387. We'll have two sets of type classes
for templates during the implementation phase of generic templates, although
the new set will be hidden from the documentation.