Generic templates: restructure type classes (#1393)
Instead of scattering the information regarding a single template across
multiple type class instances, we put it all into a single type class,
more precisley into its default method implementations.

This has the additional benefit that we can merge the `Creatable` class
into `Template` and `Exercisable` into `Choice` again.

The only potential downside of this approach is that we can call, say,
`signatory` on `Proposal t` only if there's an instance
`ProposalInstance t`. However, this might actually also be an upside.

This is part of #1387.