Switch to using haskell-lsp for IO
I’ve tried to keep this as small as possible for review so this PR
does not attempt to remove a lot of the things that are now unused.

There is also still a compatibility layer in place to avoid changing
too many things at once.

I tested this in the IDE on Linux and Windows and diagnostics,
scenario results, hover, goto definition all seem to work as expected.