Stop sending daml/keepAlive notifications
These notifications appear to be completely useless. We do have a
keepAlive mechanism in the extension but it is based on the extension
sending a daml/keepAlive request to which we then respond using a
daml/keepAlive response. This is unchanged by this PR. The
daml/keepAlive notifications however, are completely unused in the
extension.