Remove empty team placeholder in CODEOWNERS file. (#1511)
The @digital-asset/empty team was added to have a placeholder in the
CODEOWNERS file to make it easy for developers to scan it and see what
to claim ownership of. It is being removed to tidy up the file with the
expectation that those who still want to claim ownership will just add
the corresponding globs themselves.