language: derive generic instances for all data declarations.
We derive generic instances for all data declarations. We will need them
to generate upgrade code. I'm aware that this might affect compile
times. If the impact is too big, we might think about adding a flag to
damlc to only derive generic instances for building but not for the IDE.