Upgrade gRPC
The main reason for this upgrade is that it might fix some segfaults
in damlc that came from the gRPC bindings but I’ve also bumped the
JVM side at the same time.

The only slightly non-trivial change is that grpc-java stopped
exporting the compiler plugin since you are now supposed to use the
high-level rules/the toolchain. However, that doesn’t quite work for
the ledger bindings since they don’t have the provider needed for
javadoc so I patched that for now (by just changing the
visibility). I’ll open an issue for that in grpc-java.

Try to switch to the latest state of the v1.21.x branch