Refactor getAssociatedVirtualResources and remove error logging (#1763)
The log statement here didn’t make any sense since 1. This is not
actually a fatal error but just a parse error or something like that
and 2. The diagnostics for that will be reported elsewhere.

The code was also unnecessarily convoluted: It duplicated the logic
for listing scenarios and it first packaged the module into a DALF only
to then try to get back the module back out.