Replace the version placeholder in extension’s package.json by the actual version
VSCode refuses to load an extension if it has a version that does not
have a semver format. Note that this always uses the version in
__VERSION__ even for daml-sdk-head but that is consistent with what we
do everything else and given that daml-sdk-head is intended for
development I don’t think there is any point in making it more
complicated than necessary.