Remove zip-archive override and upgrade stackage snapshot
The override of zip-archive was required at some point because the new
version added additional safety checks that blew up in our tests. The
codepaths affected by this are long gone so let’s just use whatever is
in stackage.

While I was touching packages anyway, I’ve also upgraded to the latest
snapshot in the LTS we are currently following.