Produce package db for single DAML-LF version in daml_package_db rule
Currently, the daml_package_db rule produces package databases for all
supported DAML-LF versions in one go. This is rather inconvient when you
want to put further files into the database that are generated using
bazel's string interpolation. However, that is exactly what I need to do
in the context of making the daml-stdlib aware of the DAML-LF version
we're compiling to via CPP.

Thus, this PR changes the daml_pacakge_db rule such that we only produce
the package database for a single DAML-LF version.