Proposed changes to APIs to get rid of wrong notion of time
The DAML stack currently is based on a few assumption which prevent its
use in actually distributed systems. The problematic assumption is

   Assuming that a single time exists

As a consequence, the daml stack currently assumes a total ordering of
events and uses this assumption in order to support crash-recovery of
applications: When a command is submitted together with a maximum record
time (MRT), the command is meant to have not succeeded if any transaction
with a record time higher than the submitted is observed.

This however only works if we assume that all transactions are totally
ordered, which only exists under the assumption of that all events can
be ordered in time.

However, by Einstein, an event is defined as a four dimensional
coordinate (x,t) and between two-events (x,t) and (x',t') it is not
generally possible to order them, unless there is a causal relationship
between both or if x = x'

Therefore, to fix the situation, this PR highlights the problematic
assumptions and proposes a workable solution. However, the changes will
break the semantics of the ledger-api, DAML and the daml-on-x stack.

However, without them, truly distributed systems like Canton or Corda
will not work with DAML, while the changes also work for systems that do
have total ordering.